rearrange functions into smaller modules, imports seem to be working
