+++ myDir.py
@@ -1,6 +1,5 @@
 ############# If you move these programs you will need to update these directories and names #############
 def project_dir(project_name):
-    dir_project = "/content/gdrive/Shareddrives/DouglasGroup/tess_check/" + project_name + "/"
+    dir_project = "/content/gdrive/My Drive/Research/tess_check/" + project_name + "/"
     return dir_project
-#############################################################################    
-    
+#############################################################################
