+++ tesscheck.py
@@ -1,56 +1,37 @@
 # this will be the main program for inspecting TESS light curves for stellar rotation
+from glob import glob
+
 
 # Import relevant modules
-#%matplotlib inline
 import numpy as np
 import pandas as pd
+import matplotlib.pylab as pylab
 import matplotlib.pyplot as plt
-#import matplotlib.cm as cm
-#import matplotlib
 import matplotlib.gridspec as gridspec
-#from astropy.visualization import astropy_mpl_style
 from glob import glob
 from astropy.io import fits
 import warnings
 warnings.filterwarnings('ignore')
-#from astroquery.vizier import Vizier
-#from astropy.coordinates import SkyCoord
-#import astropy.units as u
 from astropy.timeseries import LombScargle
+
+# Used for tess_inspect_not_working - probably not needed anymore?
+import ipywidgets as widgets
+from ipywidgets import interactive
+from IPython.display import display
 
 import os
 import sys
 ############# If you move these programs you will need to update these directories and names #############
 sys.path.append('/content/gdrive/My Drive/')
-from tess_check import myDir as myDir
+from . import myDir as myDir
+from .status import read_status
+from .lightcurves import find_sap, find_cpm
+from .lightcurves import load_cpm, load_cpm_fromfile, load_sap
+from .ffis import find_ffi, load_ffi_fromfile, ffi_test
+from .sheet import get_prot_table
 
 import time
 ####################################################################################
-
-# Load sheet
-def load_sheet(project_name):
-    from google.colab import auth
-    auth.authenticate_user()
-    import gspread
-    from oauth2client.client import GoogleCredentials
-    gc = gspread.authorize(GoogleCredentials.get_application_default())
-
-    dir_project = myDir.project_dir(project_name)
-    sheet_id_file = os.path.join(dir_project,f"{project_name}.txt")
-
-    f = open(sheet_id_file,"r")
-    doc_id = f.readline()
-    f.close()
-    sheet = gc.open_by_key(doc_id)
-    return sheet
-
-
-def list_to_string(list,delim=' '):
-  list_as_string = list[0]
-  for i in range(len(list)-1):
-    id = i+1
-    list_as_string += (delim + list[id])
-  return list_as_string
 
 # Auto run
 def tesscheck_auto(project_name, tess_cycle=1, redo=False):
@@ -92,7 +73,7 @@ def tesscheck_run_v1():
 
 	# Load the Prot table
 	prot_table = get_prot_table(user)
-	
+
 #	rows = prot_table.get_all_values()
 #	print(rows)
 
@@ -107,171 +88,21 @@ def tesscheck_run_v1():
 	star = clu.iloc[276]
 	tstar = initiate_star(star,cluster,user=user)
 	#display_tess_lite(tstar, save = True)
-	
+
 	display_tess_lite_v2(tstar, save = False, noplot = True)
-	
+
 	update_prot_table(prot_table, tstar, user)
-	
+
 	return tstar
 
 
-####################################################################################
-# Identify the User
-def tess_user(project_name):
-    #dir_project = project_dir(project_name)
-    status = read_status(project_name)
 
-    #file_status = glob(dir_project + 'Status.txt')
-    #file_open = open(file_status[0], "r")
-    #lines = file_open.readlines()
-    #user_line = lines[8]
-    users = status['Users'].split(' ')
-    #users = users[1:-1]
-    number_of_users = len(users)
-    
-    print('Which user? Press...')
-    for i in range(number_of_users):
-        print('   '+str(i+1)+' for '+users[i])
-    print('   '+str(i+2)+' for Other')
-
-# while loop
-
-#  val = input("Enter number for your name: ") 
-    val = input()
-    user = ''
-    #user_found = 0
-    
-    if val.isdigit() == False:
-        print('No user selected.')
-        user = 'None'
-        return user
-    else: 
-    # is the number in the range?
-        if (float(val) > (number_of_users+1)) | (float(val) == 0):
-            print('Out of bounds')
-            user = 'Noone'
-            return user
-        if (float(val) <= (number_of_users)) & (float(val) != 0):
-            id = int(val)
-            user = users[id-1]
-            print(user + ' is logged in.')
-            add_user_to_sheet(project_name, user)
-            return user
-        if float(val) == (number_of_users+1):
-            print('Other selected. Need to make sheet and update Status.txt')
-            print('Other: What name?')
-            other_name = input()
-            other_name = other_name.replace(" ", "")
-            other_name = other_name.lower()
-            other_name = other_name.capitalize()
-            #make_status(project_name, add_user=other_name, add_step = '', change_auto='', reset=False):
-            user = other_name # prompt for it
-            iu = np.where(np.array(users) == user)
-            if np.size(iu) > 0: 
-                print('User already exists, logging in.')
-                add_user_to_sheet(project_name, user)
-                return user
-            else:
-                make_status(project_name, add_user=user, add_step = '', change_auto='', reset=False)
-                add_user_to_sheet(project_name, user)
-                print(user + ' profile is created and user is logged in.')
-            return user
-
-####################################################################################
-def get_prot_table(user,project_name):
-	sheet_name = project_name
-	from google.colab import auth
-	auth.authenticate_user()
-	import gspread
-	from oauth2client.client import GoogleCredentials
-	gc = gspread.authorize(GoogleCredentials.get_application_default())
-	# load the table
-	worksheet = load_sheet(sheet_name)
-	if user == 'Auto':
-		table = worksheet.worksheet('Targets')
-	else:
-		table = worksheet.worksheet(user)
-	return table
-####################################################################################
-# Functions to locate a star's TESS data
-
-def find_ffi(star, cluster):
-#    homedir = os.path.expanduser("~")
-    dir_ffi = myDir.project_dir(cluster)+'FFI/'
-    RA = str(star["RA_ICRS"])[:7]
-    DE = str(star["DE_ICRS"])[:7]
-    file_ffi = glob(dir_ffi+"*"+RA+"*"+DE+"*/*.fits")
-    if len(file_ffi) == 0:
-        file_ffi = glob(dir_ffi+"*"+RA+"*"+DE+"*.fits")
-    return(file_ffi)
-
-def find_ffi_coord(ra, dec, cluster):
-    dir_ffi = myDir.project_dir(cluster)+'FFI/'
-    RA = str(ra)[:7]
-    DE = str(dec)[:7]
-    file_ffi = glob(dir_ffi+"*"+RA+"*"+DE+"*/*.fits")
-    if len(file_ffi) == 0:
-        file_ffi = glob(dir_ffi+"*"+RA+"*"+DE+"*.fits")
-    return(file_ffi)
-
-
-def find_sap(star, cluster):
-    dir_sap = myDir.project_dir(cluster)+'SAP/'
-    if star["DR2NAME"][0] == 'G':
-        star_name = star["DR2NAME"][9:]
-    else:
-        star_name = star["DR2NAME"]
-    file_sap = glob(dir_sap + "*" + star_name + "*.csv")
-    return(file_sap)
-
-def find_cpm(star, cluster):
-    dir_cpm = myDir.project_dir(cluster)+'CPM/'
-    if star["DR2NAME"][0] == 'G':
-        star_name = star["DR2NAME"][9:]
-    else:
-        star_name = star["DR2NAME"]
-    file_cpm = glob(dir_cpm + "*" + star_name + "*.csv")
-    return(file_cpm)
-
-def load_cpm_fromfile(file):
-    lc = pd.read_csv(file)
-    return lc
-
-def load_cpm(star):
-    lc = pd.read_csv(star['file_cpm'][0])
-    return lc
-
-def load_sap(star):
-    lc = pd.read_csv(star['file_sap'][0])
-    return lc
-
-def load_ffi_fromfile(file):
-    ffi_data = fits.open(file)
-    images = ffi_data[1].data
-    image = images[100]['Flux']
-    if np.max(image) == 0:
-      image = images[500]['Flux']
-    return image
-
-def load_ffi(star):
-    ffi_data = fits.open(star['file_ffi'][0])
-    images = ffi_data[1].data
-    image = images[100]['Flux']
-    if np.max(image) == 0:
-      image = images[500]['Flux']
-    return image
-
-def load_ffis(star):
-    ffi_data = fits.open(star['file_ffi'][0])
-    images = ffi_data[1].data
-    image = images['Flux']
-    return image
 ####################################################################################
 def make_star(target_data, dr2_now):
     star = {'DR2NAME': '',
             'RA_ICRS': 0.,
             'DE_ICRS': 0.,
-            'GMAG': 0., 
+            'GMAG': 0.,
             'BP_RP': 0.}
 
     iloc = np.where(dr2_now == target_data['DR2Name'])
@@ -310,7 +141,7 @@ def initiate_star(star, cluster, user='NONE', blank=False):
                 'Flares':0,
                 'Notes':'', # anything noteworthy about this star?
                 'LC_Quality':1, # 1 = modulate, 0 is flat, -1 is garbage
-                'LC_Action':'', # 
+                'LC_Action':'', #
                 # Plotting options
                 'x_min':0.0,'x_max':0.0, # time range
                 'y_min':0.0,'y_max':0.0, # flux range, will be calculated during first iteration, and then adjusted by user
@@ -343,7 +174,7 @@ def initiate_star(star, cluster, user='NONE', blank=False):
     if len(file_cpm) > 0:
         exist[1] = 1
         #lc_cpm = load_cpm(file_cpm)
-        star_data['file_cpm'] = file_cpm            
+        star_data['file_cpm'] = file_cpm
         star_data['exist_cpm'] = 1
     if len(file_sap) > 0:
         exist[2] = 1
@@ -366,14 +197,13 @@ def display_tess_lite_v2(tstar,save = False,noplot = False):
     time_1 = time.time()
     # plotting defaults
     axis_fontsize = 16
-    import matplotlib.pylab as pylab
     params = {'axes.labelsize': 16,'axes.titlesize': 16,'xtick.labelsize': 14,'ytick.labelsize': 14}
     pylab.rcParams.update(params)
-    
+
 #cpm data for star
     if tstar['exist_cpm'] == 0:
         return
-    
+
     if tstar['which_LC'] == 'CPM':
         lc_cpm = load_cpm(tstar)
 #        if tstar['which_sectors'] != 'All':
@@ -399,7 +229,7 @@ def display_tess_lite_v2(tstar,save = False,noplot = False):
     ifin = np.where(np.isfinite(flux_all)) #find infs
     unique_sectors = np.unique(sector_all).astype(int) #all the sectors for each data point from table
     unique_sectors = list(map(str,unique_sectors)) #unique instances
-    # save into 
+    # save into
     tstar['sector_list'] = unique_sectors #unique sectors saved as a list
     length_all = len(flux_all)
     use_these = np.zeros(length_all)
@@ -440,7 +270,7 @@ def display_tess_lite_v2(tstar,save = False,noplot = False):
     times = time_all[iuse[0]]
     flux = flux_all[iuse[0]]
     sectors = sector_all[iuse[0]]
-    sectors_used = np.unique(sectors).astype(int)     
+    sectors_used = np.unique(sectors).astype(int)
     sectors_used = list(map(str,sectors_used))
 
 
@@ -469,7 +299,7 @@ def display_tess_lite_v2(tstar,save = False,noplot = False):
     period = periods_cpm[np.argmax(pow_cpm)]
     tstar['Prot_LS'] = period
     tstar['Power_LS'] = np.max(pow_cpm)
-    
+
 # Amplitude measurement
 
     perc05 = np.percentile(flux,5)
@@ -479,29 +309,29 @@ def display_tess_lite_v2(tstar,save = False,noplot = False):
 
 
 # check if double
-# read which_LC, then store in that period. 
+# read which_LC, then store in that period.
 # store these in star, update Prot_final
 
-    mdub = float(1.0) 
+    mdub = float(1.0)
     if tstar['is_it_double'] == 1:
         mdub = float(2.0)
     period_final = float(tstar['Prot_LS']*mdub)
     tstar['Prot_final'] = period_final
-    
-   
+
+
 #Figure creation
     if noplot == False:
         panel = plt.figure(constrained_layout=True, figsize= (16,11))
         gs = gridspec.GridSpec(100, 100)
 
-#cpm lightcurve 
+#cpm lightcurve
 # how many sectors?
     #all_sectors = lc_cpm['sector'].unique().astype(int)
 #    unique_sectors = sector_all.unique().astype(int)
 #    all_sectors = sectors # I'm pretty sure I reran CPM so that this isn't an issue anymore. Bad sectors arent in the CPM file.
 
     n_sec = len(sectors_used) # this should probably be number used
-    n_all_sec = len(unique_sectors)  
+    n_all_sec = len(unique_sectors)
     tstar['number_sectors'] = n_sec
     primary_colors = ['b','r']
     color_options = []
@@ -532,7 +362,7 @@ def display_tess_lite_v2(tstar,save = False,noplot = False):
 
         #Mark adding a text for each sector as it is plotted
         bot, top = cpmlight.get_ylim() #find the upper limit for us to put text
-        
+
         for x in sectors.index:
             if x == sectors.index.min():
                 cpmlight.text(times[x]-tmin, top*.9, str(int(sectors[x]))) #put sector number for first sector
@@ -571,8 +401,8 @@ def display_tess_lite_v2(tstar,save = False,noplot = False):
 
 #  print(cpmlight.get_xlim())
 
-#    print('First panels: '+str(time.time()-time_3))  	
-#   First panels: 0.05 seconds    
+#    print('First panels: '+str(time.time()-time_3))
+#   First panels: 0.05 seconds
 
 #FFI image
 
@@ -595,10 +425,10 @@ def display_tess_lite_v2(tstar,save = False,noplot = False):
         reversed_color_map = color_map.reversed()
         ffi_mod = np.clip(ffi_image-np.min(ffi_image),0,1000)
         ffimage.imshow(ffi_mod,origin = 'lower',cmap=reversed_color_map)
-        ffimage.plot([15,17],[20,20],color='red')    
-        ffimage.plot([23,25],[20,20],color='red')    
-        ffimage.plot([20,20],[15,17],color='red')    
-        ffimage.plot([20,20],[23,25],color='red')    
+        ffimage.plot([15,17],[20,20],color='red')
+        ffimage.plot([23,25],[20,20],color='red')
+        ffimage.plot([20,20],[15,17],color='red')
+        ffimage.plot([20,20],[23,25],color='red')
         ffimage.set_xlabel('Pixels')
         ffimage.set_ylabel('Pixels')
 
@@ -633,11 +463,11 @@ def update_prot_table(table, tstar):
         n_col = len(columns)
         cols = []
         for column in columns:
-            cell = table.find(column)    
+            cell = table.find(column)
             cols.append(cell.col)
         cell_list = [table.cell(row_number,cols[0]),table.cell(row_number,cols[1]),table.cell(row_number,cols[2]),table.cell(row_number,cols[3])]
-        cell_list[1].value = tstar['Prot_LS']            
-        cell_list[2].value = tstar['Power_LS']            
+        cell_list[1].value = tstar['Prot_LS']
+        cell_list[2].value = tstar['Power_LS']
         if (tstar['exist_ffi'] == 0) or (tstar['exist_cpm'] == 0):
             cell_list[0].value = '-1'
             cell_list[3].value = 'No'
@@ -646,13 +476,13 @@ def update_prot_table(table, tstar):
             cell_list[3].value = 'Yes'
         table.update_cells(cell_list)
 
-        #added amplitude column, and sector_list 
+        #added amplitude column, and sector_list
     if user != 'Auto':
         columns = ['Prot_Final','Prot_LS', 'Power_LS', 'Single_Double', 'Multi', 'Quality', 'LC_Source', 'Class', 'Notes', 'Amp','Sectors_Used','Flares']
         n_col = len(columns)
         cols = [2,3,4,5,6,7,8,9,10,11,12,13]
 #        for column in columns:
-#            cell = table.find(column)    
+#            cell = table.find(column)
 #            cols.append(cell.col)
         cell_range = 'B'+str(row_number)+':M'+str(row_number)
         cell_list = table.range(cell_range)
@@ -666,124 +496,24 @@ def update_prot_table(table, tstar):
             cell_list[0].value = 99
         if tstar['LC_Action'] == 'Garbage':
             cell_list[0].value = -99
-        cell_list[1].value = tstar['Prot_LS']            
-        cell_list[2].value = tstar['Power_LS']            
-        cell_list[3].value = tstar['is_it_double']            
-        cell_list[4].value = tstar['Multi']            
-        cell_list[5].value = tstar['LC_Quality']            
-        cell_list[6].value = tstar['which_LC']            
-        cell_list[7].value = tstar['LC_Action']                    
+        cell_list[1].value = tstar['Prot_LS']
+        cell_list[2].value = tstar['Power_LS']
+        cell_list[3].value = tstar['is_it_double']
+        cell_list[4].value = tstar['Multi']
+        cell_list[5].value = tstar['LC_Quality']
+        cell_list[6].value = tstar['which_LC']
+        cell_list[7].value = tstar['LC_Action']
         cell_list[8].value = tstar['Notes']
         cell_list[9].value = tstar['Amplitude']
         cell_list[10].value = str(tstar['which_sectors'])
         cell_list[11].value = tstar['Flares']
         table.update_cells(cell_list)
 
-####################################################################################
-def stars_todo(table):
-    n_stars = len(table)
-    not_done = np.zeros(n_stars, dtype=int)
-    for i in range(n_stars):
-        if len(table['Prot_LS'][i]) == 0:
-            not_done[i] = 1
-            
-    ido = np.where(not_done == 1)
-    dr2_list = table['DR2Name'].to_numpy()
-    star_list = dr2_list[ido[0]]
-
-    return star_list 
-
-def stars_todo_split(table, user):
-    n_stars = len(table)
-    not_done = np.zeros(n_stars, dtype=int)
-    for i in range(n_stars):
-        if len(table['Prot_LS'][i]) == 0:
-            not_done[i] = 1
-            
-    ido = np.where(not_done == 1)
-    list = ido[0]
-    if user == 'Angeli':
-        these = np.where(ido[0] < 202)
-        list = ido[0][these]
-    if user == 'Isabella':
-        these = np.where((ido[0] > 191) & (ido[0] < 379))
-        list = ido[0][these]
-    if user == 'Linus':
-        these = np.where(ido[0] > 373)
-        list = ido[0][these]
-
-    dr2_list = table['DR2Name'].to_numpy()
-    star_list = dr2_list[list]
-
-    return star_list 
-
-####################################################################################    
-def stars_nodata():
-    target_table = get_prot_table('Auto')
-    target_data = target_table.get_all_records()
-    dr2_list = target_table.col_values(1)
-    dr2 = np.array(dr2_list[1:])
-    num = np.size(dr2)
-
-# load their table
-    user_table = get_prot_table('Jason')
-# Identify columns        
-# Locate Prot_LS column
-    cell = user_table.find('Prot_LS')
-    col_prot = cell.col
-# Locate Power LS column
-    cell = user_table.find('Power_LS')
-    col_pow = cell.col
-# Loop over targets
-    for i in range(num):
-        row_number = i+2
-# Does target have TESS data, according to target table?
-        val = target_data[i]['Prot_LS']
-        if val == 0:
-            user_table.update_cell(row_number,col_prot,0)
-            user_table.update_cell(row_number,col_pow,0)
-####################################################################################
-def stars_nodata_new(Team):
-    target_table = get_prot_table('Auto')
-    target_data = target_table.get_all_records()
-    dr2_list = target_table.col_values(1)
-    dr2 = np.array(dr2_list[1:])
-    num = np.size(dr2)
-
-# load their table
-    user_table = get_prot_table('Jason')
-# Identify columns        
-# Locate Prot_LS column
-    cell = user_table.find('Prot_LS')
-    col_prot = cell.col
-# Locate Power LS column
-    cell = user_table.find('Power_LS')
-    col_pow = cell.col
-# Loop over targets
-    for i in range(num):
-        row_number = i+2
-# Does target have TESS data, according to target table?
-        val = target_data[i]['Prot_LS']
-        if val == 0:
-            user_table.update_cell(row_number,col_prot,0)
-            user_table.update_cell(row_number,col_pow,0)
 
 ####################################################################################
-def ffi_test(ffi):
-    shape = np.shape(ffi)
-    val = ffi[int(shape[0]/2),int(shape[1]/2)]
-    if np.isfinite(val):
-      good_or_bad = 1
-    else:
-      good_or_bad = 0
-    return good_or_bad
-####################################################################################
+# probably not needed anymore?
 def tess_inspect_not_working(tstar):
-    import ipywidgets as widgets
-    from ipywidgets import interactive
-    from IPython.display import display
-    import matplotlib.pylab as pylab
-  
+
     lc_cpm = load_cpm(tstar)
     all_sectors = lc_cpm['sector'].unique().astype(int)
     sectors = sector.unique().astype(int)
@@ -804,8 +534,6 @@ def tess_inspect_not_working(tstar):
 ####################################################################################
 ####################################################################################
 def update_panelname_v1(tstar, locate=False):
-    import os
-    from glob import glob
     dir_panels = '/content/gdrive/My Drive/Projects/'+tstar['Cluster']+'/Panels/'
     name = dir_panels + '*'+tstar['Source'][9:]+'*'
     file = glob(name)
@@ -815,10 +543,8 @@ def update_panelname_v1(tstar, locate=False):
         end = '-User='+tstar['User']+'-Review='+tstar['LC_Action']+'.png'
         new_file = dir_panels + 'GaiaDR2_'+str(tstar["Source"])[9:]+end
         os.rename(file[0],new_file)
-        
+
 def update_panelname(tstar, locate=False):
-    import os
-    from glob import glob
     dir_panels = myDir.project_dir(tstar['Cluster'])+'Panels/'
     if tstar['Source'][0] == 'G':
         dr2 = tstar['Source'][9:]
@@ -831,51 +557,12 @@ def update_panelname(tstar, locate=False):
     else:
         end = '-User='+tstar['User']+'-Review='+tstar['LC_Action']+'.png'
         new_file = dir_panels + 'GaiaDR2_'+dr2+end
-        os.rename(file[0],new_file)        
+        os.rename(file[0],new_file)
 
-#############################################################################
-def prot_show_v1(project_name, user, gbr, clusters=False):
-#    gbr = target_data['BP_RP'].to_numpy(dtype=float)
-    fig1, ax1 = plt.subplots(figsize=(7,6)) 
-    ax1.tick_params(axis='both', which='major', labelsize=15)
-    aw = 1.5
-    ax1.spines['top'].set_linewidth(aw)
-    ax1.spines['left'].set_linewidth(aw)
-    ax1.spines['right'].set_linewidth(aw)
-    ax1.spines['bottom'].set_linewidth(aw)
 
-    prot_table_now = get_prot_table(user,project_name)
-    prot_data_val_now = prot_table_now.get_all_values()
-    prot_data_now = pd.DataFrame.from_records(prot_data_val_now[1:],columns=prot_data_val_now[0])
-    
-    pnow = prot_data_now['Prot_Final'].to_numpy()
-    uu = np.where((pnow != '') & (pnow != '-1') & (gbr != 'nan'))
-    prot_now = pnow[uu[0]]
-    color = gbr[uu[0]].astype(float)
-    ax1.set_xlim(0.4,2.5)
-    ax1.set_xlabel('BP - RP (mag)',fontsize=20)
-    ax1.set_ylim(0,20)
-    ax1.set_ylabel('Rotation Period (days)',fontsize=20)
-    
-    if clusters == True:
-        file = glob('/content/gdrive/My Drive/Tables/gyro_clusters_draft-2020April08.csv')
-        clus = pd.read_csv(file[0])
-        indicesPl = np.where((clus["CLUSTER"] ==  "Pleiades") & (clus['BENCH'] == 1))
-        indicesPr = np.where((clus["CLUSTER"] == "Praesepe") & (clus['BENCH'] == 1))
-        #indicesNGC = np.where((Cluster == "NGC_6811") & (clus['BENCH'] == 1))
-        pleiades = clus.iloc[indicesPl]
-        praesepe = clus.iloc[indicesPr]
-        #NGC6811 = clus.iloc[indicesNGC]
-        plt.plot(pleiades["BP_RP"]-0.415*0.12, pleiades["PROT"], markerfacecolor = 'blue', markeredgecolor='black', label = '120 Myr Pleiades',markersize=10,alpha=0.7,linestyle='',marker='.')
-        plt.plot(praesepe["BP_RP"]-0.415*0.035, praesepe["PROT"], markerfacecolor = 'cyan', markeredgecolor='black', label = '670 Myr Praesepe',markersize=10,alpha=0.7,linestyle='',marker='.')
-
-    ax1.plot(color, np.array(prot_now,dtype=float),markerfacecolor='red',markeredgecolor='black',marker='*',markersize=20,linestyle='')
-    plt.show()
-    
-    
 def prot_show(project_name, user, gbr, clusters=False, pcut=0.0):
 #    gbr = target_data['BP_RP'].to_numpy(dtype=float)
-    fig1, ax1 = plt.subplots(figsize=(15,9)) 
+    fig1, ax1 = plt.subplots(figsize=(15,9))
     ax1.tick_params(axis='both', which='major', labelsize=15)
     aw = 1.5
     ax1.spines['top'].set_linewidth(aw)
@@ -886,7 +573,7 @@ def prot_show(project_name, user, gbr, clusters=False, pcut=0.0):
     prot_table_now = get_prot_table(user,project_name)
     prot_data_val_now = prot_table_now.get_all_values()
     prot_data_now = pd.DataFrame.from_records(prot_data_val_now[1:],columns=prot_data_val_now[0])
-    
+
     pnow = prot_data_now['Prot_Final'].to_numpy()
     qnow = prot_data_now['Quality'].to_numpy()
     cnow = prot_data_now['Class'].to_numpy()
@@ -902,7 +589,7 @@ def prot_show(project_name, user, gbr, clusters=False, pcut=0.0):
     ax1.set_ylim(0,25)
     ax1.set_ylabel('Rotation Period (days)',fontsize=20)
 
-    
+
     if clusters == True:
         file = glob('/content/gdrive/My Drive/Tables/gyro_clusters_draft-2020April08.csv')
         clus = pd.read_csv(file[0])
@@ -922,173 +609,9 @@ def prot_show(project_name, user, gbr, clusters=False, pcut=0.0):
 #    ax1.plot([1.2375,1.2375],[0,20],c='green')
 #    ax1.plot([0.5,2.5],[11.677,11.677],c='green')
     plt.show()
-    
-#############################################################################
-def make_status(project_name, add_user='', add_step = '', change_auto='', reset=False):
-    # directory
-    dir_project = myDir.project_dir(project_name)
-    # ensure the file doesnt already exist
-    file_status = glob(dir_project + "Status.txt")
-
-    new_file = 1
-    if (np.size(file_status) == 1) | (reset == False):
-        bsize = os.path.getsize(file_status[0]) #size in bytes
-        if bsize < 40:
-            print('remove the file')
-            os.remove(file_status[0])  
-        else:
-            new_file = 0
-            status = read_status(project_name)
-
-    if (new_file == 1) | (reset == True):
-        status = {'Project':project_name,
-            'Users':'Jason Team_Member',
-            'Steps':'Status_Initialized',
-            'Auto': 'No'}
-
-    if len(add_user) > 0:
-        status['Users'] += ' '+add_user
-
-    if len(add_step) > 0:
-        status['Steps'] += ' '+add_step
-
-    if len(change_auto) > 0:
-        status['Auto'] = change_auto
-
-    lines = []
-    # 1: project title
-    lines.append('Project: '+project_name+"\n")
-    # 2: Users
-    lines.append('Users: '+status['Users']+"\n")
-    # 3: Steps
-    lines.append('Steps: '+status['Steps']+"\n")
-    # 4: Has tesscheck_auto been run?
-    lines.append('tesscheck_auto: '+status['Auto']+"")
-    # 5: Which sectors?
-    # 6: Number of repeat sectors?
-
-    # Create the file
-    fo = open(dir_project + "Status.txt", "w")
-    fo.writelines(lines)
-    fo.close()
-######################################
-def read_status(project_name):
-    dir_project = myDir.project_dir(project_name)
-    # ensure the file doesnt already exist
-    file_status = glob(dir_project + "Status.txt")
-    if np.size(file_status) == 0:
-        make_status(project_name)
-        #print('no file')
-        return
-
-    bsize = os.path.getsize(file_status[0]) #size in bytes
-    if (bsize < 40):
-        os.remove(file_status[0])  
-        make_status(project_name)
-        #print('no file')
-        return
-
-    fo = open(file_status[0], "r")
-    lines_current = fo.readlines()
-    fo.close()
-
-    # 2 Users
-    c_users_line = lines_current[1][0:-1]
-    c_users_split = c_users_line.split(sep=' ')
-    users_current = c_users_split[1:]
-
-    # 3 Steps
-    c_steps_line = lines_current[2][0:-1]
-    c_steps_split = c_steps_line.split(sep=' ')
-    steps_current = c_steps_split[1:]
-
-    # 4 Auto
-    c_line = lines_current[3]
-    c_split = c_line.split(sep=' ')
-    auto_current = c_split[1:]
-
-    status = {'Project':project_name,
-            'Users':list_to_string(users_current),
-            'Steps':list_to_string(steps_current),
-            'Auto':auto_current[0]}
-            
-    return status
-#############################################################################
-def add_user_to_sheet(project_name, user):
-    from google.colab import auth
-    auth.authenticate_user()
-    import gspread
-    from oauth2client.client import GoogleCredentials
-    gc = gspread.authorize(GoogleCredentials.get_application_default())
-    sheet = load_sheet(project_name)
-    sheet_list = sheet.worksheets()
-
-    target_table = get_prot_table('Auto',project_name)
-    target_data_val = target_table.get_all_values()
-    target_data = pd.DataFrame.from_records(target_data_val[1:],columns=target_data_val[0])
-    dr2_list = target_data['DR2Name'].to_numpy()
-    
-    number_of_stars = len(dr2_list)
-
-    sheet_exists = 0
-    for isheet in sheet_list:
-        if isheet.title == user:
-            sheet_exists = 1
-
-    if sheet_exists == 1:
-        print('This user has a sheet')
-    else:
-        print('Making sheet for new user...')
-        new_sheet = sheet.add_worksheet(title=user,rows=number_of_stars+1,cols=10)
-        columns = ['DR2Name', 'Prot_Final', 'Prot_LS', 'Power_LS', 'Single_Double', 'Multi', 'Quality', 'LC_Source', 'Class', 'Notes','Amp','Sectors_Used','Flares']
-        cell_range = 'A1:M1'
-        cell_list = new_sheet.range(cell_range)
-        i=0
-        for cell in cell_list:
-            cell.value = columns[i]
-            i+=1
-        new_sheet.update_cells(cell_list)
-
-        cell_range = 'A2:A'+str(number_of_stars+1)
-        cell_list = new_sheet.range(cell_range)
-        i=0
-        for cell in cell_list:
-            cell.value = target_data['DR2Name'][i]
-            i+=1
-        new_sheet.update_cells(cell_list)
-
-        cell_range = 'B2:B'+str(number_of_stars+1)
-        cell_list = new_sheet.range(cell_range)
-        i=0
-        for cell in cell_list:
-            if target_data['Prot'][i] == '-1':
-                cell.value = target_data['Prot'][i]
-            i+=1
-        new_sheet.update_cells(cell_list)
-
-        cell_range = 'C2:C'+str(number_of_stars+1)
-        cell_list = new_sheet.range(cell_range)
-        i=0
-        for cell in cell_list:
-            if target_data['Prot_LS'][i] == '0':
-                cell.value = target_data['Prot_LS'][i]
-            i+=1
-        new_sheet.update_cells(cell_list)
-
-        cell_range = 'D2:D'+str(number_of_stars+1)
-        cell_list = new_sheet.range(cell_range)
-        i=0
-        for cell in cell_list:
-            if target_data['Power_LS'][i] == '0':
-                cell.value = target_data['Power_LS'][i]
-            i+=1
-        new_sheet.update_cells(cell_list)
-    
-#    print('Setting up a new user took '+str(time.time()-start_time)+' seconds')
-
 
 def prot_auto_show(project_name, clusters=False, pcut=0.0, av=0.0):
-    fig1, ax1 = plt.subplots(figsize=(15,9)) 
+    fig1, ax1 = plt.subplots(figsize=(15,9))
     ax1.tick_params(axis='both', which='major', labelsize=15)
     aw = 1.5
     ax1.spines['top'].set_linewidth(aw)
@@ -1099,10 +622,10 @@ def prot_auto_show(project_name, clusters=False, pcut=0.0, av=0.0):
     worksheet = load_sheet(project_name)
     target_table = worksheet.worksheet('Targets')
     target_data_val = target_table.get_all_values()
-    target_data = pd.DataFrame.from_records(target_data_val[1:],columns=target_data_val[0])    
-    
+    target_data = pd.DataFrame.from_records(target_data_val[1:],columns=target_data_val[0])
+
     gbr = target_data['BP_RP'].to_numpy(dtype=float)
-    
+
     pnow = target_data['Prot_LS'].to_numpy()
     uu = np.where((pnow != '') & (pnow != '-1') & (gbr != 'nan'))
     prot_now = pnow[uu[0]]
@@ -1114,7 +637,7 @@ def prot_auto_show(project_name, clusters=False, pcut=0.0, av=0.0):
     ax1.set_ylim(0,25)
     ax1.set_ylabel('Rotation Period (days)',fontsize=20)
 
-    
+
     if clusters == True:
         file = glob('/content/gdrive/My Drive/Tables/gyro_clusters_draft-2020April08.csv')
         clus = pd.read_csv(file[0])
@@ -1135,4 +658,3 @@ def prot_auto_show(project_name, clusters=False, pcut=0.0, av=0.0):
 #    ax1.plot([1.2375,1.2375],[0,20],c='green')
 #    ax1.plot([0.5,2.5],[11.677,11.677],c='green')
     plt.show()
-
